% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{auckland_islands_nz}
\alias{auckland_islands_nz}
\title{auckland_islands_nz}
\format{
A .rda file containing the objects:
\describe{
  \item{auckland_islands_nz}{Random seed and bounding box for the mastersample for the Auckland Islands of New Zealand}
}
}
\usage{
data(auckland_islands_nz)
}
\description{
This dataset contains seeds and bounding boxes
for the mastersamples accessed from spbal: `auckland_islands_nz`.
}
\keyword{datasets}
