% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{bc_ca}
\alias{bc_ca}
\title{bc_ca}
\format{
A .rda file containing the objects:
\describe{
  \item{bc_ca}{Random seed and bounding box for the mastersample for the Auckland Islands of New Zealand}
}
}
\usage{
data(bc_ca)
}
\description{
This dataset contains seeds and bounding boxes
for the mastersamples accessed from spbal:`bc_ca`.
}
\keyword{datasets}
