% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{north_nz}
\alias{north_nz}
\title{north_nz}
\format{
A .rda file containing the objects:
\describe{
  \item{north_nz}{Random seed and bounding box for the mastersample for the North Island of New Zealand}
}
}
\usage{
data(north_nz)
}
\description{
This dataset contains seeds and bounding boxes
for the mastersamples accessed from spbal: `north_nz`.
}
\keyword{datasets}
