% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{south_nz}
\alias{south_nz}
\title{south_nz}
\format{
A .rda file containing the objects:
\describe{
  \item{south_nz}{Random seed and bounding box for the mastersample for the South Island of New Zealand}
}
}
\usage{
data(south_nz)
}
\description{
This dataset contains seeds and bounding boxes
for the mastersamples accessed from spbal: `south_nz`.
}
\keyword{datasets}
