% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{uniqueID}
\alias{uniqueID}
\title{uniqueID}
\usage{
uniqueID(mastersample, sample)
}
\arguments{
\item{mastersample}{The existing mastersample to use}

\item{sample}{The sample drawn from spbal}
}
\value{
An sf point file from spbal with the uniqueID of an existing mastersample appended to it
}
\description{
Append a mastersample uniqueID to a sample draw from spbal
}
\examples{
#Append the uniqueID to a sample draw from the mastersample for the South Island, New Zealand
uniqueID(south_nz,BASsample)

}
